refactor: move files to appropriate modules
