+++ src/cms/types/contentTypeSchema.types.ts
