+++ src/index.ts
@@ -1,4 +1,4 @@
-import ContentstackLivePreviewHOC from "./contentstack-live-preview-HOC";
+import ContentstackLivePreviewHOC from "./preview/contentstack-live-preview-HOC";
 import "./styles.css";
 
 export const ContentstackLivePreview = ContentstackLivePreviewHOC;
