+++ src/liveEditor/utils/multipleElementAddButton.ts
@@ -1,4 +1,4 @@
-import { VisualEditorCslpEventDetails } from "../../types/liveEditor.types";
+import { VisualEditorCslpEventDetails } from "../types/liveEditor.types";
 import { FieldSchemaMap } from "./fieldSchemaMap";
 import {
     generateAddInstanceButton,
