+++ src/liveEditor/utils/pseudoEditableField.ts
@@ -1,5 +1,5 @@
 import { toString } from "lodash";
-import { CslpData } from "../../types/cslp.types";
+import { CslpData } from "../../cslp/types/cslp.types";
 import liveEditorPostMessage from "./liveEditorPostMessage";
 import { LiveEditorPostMessageEvents } from "./types/postMessage.types";
