+++ tsconfig.json
@@ -74,5 +74,5 @@
         "skipLibCheck": true /* Skip type checking of declaration files. */,
         "forceConsistentCasingInFileNames": true /* Disallow inconsistently-cased references to the same file. */
     },
-    "exclude": ["node_modules", "**/__test__", "dist"]
+    "exclude": ["node_modules", "dist"]
 }
